#7 chore: 로그인/회원가입페이지 로고  Image from 'next/image' 추가
