+++ src/app/signup/page.tsx
@@ -2,6 +2,7 @@
 
 import signupAPI from '@/api/auth/signupAPI';
 import { useState } from 'react';
+import Image from 'next/image';
 
 const SignupPage = () => {
   const [email, setEmail] = useState('');
@@ -10,7 +11,7 @@ const SignupPage = () => {
   const [message, setMessage] = useState('');
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault(); // <- 짜증나는 놈 ㅠ_ㅠ
+    event.preventDefault();
 
     try {
       const response = await signupAPI({
@@ -29,24 +30,27 @@ const SignupPage = () => {
     }
   };
   return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label>이메일</label>
-          <input type='email' value={email} onChange={(e) => setEmail(e.target.value)} />
-        </div>
-        <div>
-          <label>닉네임</label>
-          <input type='text' value={nickname} onChange={(e) => setNickname(e.target.value)} />
-        </div>
-        <div>
-          <label>비밀번호</label>
-          <input type='password' value={password} onChange={(e) => setPassword(e.target.value)} />
-        </div>
-        <button type='submit'>가입하기</button>
-      </form>
-      {message && <p>{message}</p>}
-    </div>
+    <>
+      <Image src='/images/images_logo.png' alt='로고 이미지' width={200} height={280} />
+      <div>
+        <form onSubmit={handleSubmit}>
+          <div>
+            <label>이메일</label>
+            <input type='email' value={email} onChange={(e) => setEmail(e.target.value)} />
+          </div>
+          <div>
+            <label>닉네임</label>
+            <input type='text' value={nickname} onChange={(e) => setNickname(e.target.value)} />
+          </div>
+          <div>
+            <label>비밀번호</label>
+            <input type='password' value={password} onChange={(e) => setPassword(e.target.value)} />
+          </div>
+          <button type='submit'>가입하기</button>
+        </form>
+        {message && <p>{message}</p>}
+      </div>
+    </>
   );
 };
